requirements.txt
aiogram==3.15.0
flake8==5.0.4
flake8-docstrings==1.6.0
pytest-timeout==2.1.0
python-dotenv==0.20.0
isort==5.13.2